+++ src/AntShares.VM/OpCode.cs
@@ -42,8 +42,8 @@
         TAILCALL = 0x69,
 
 
-        DUPFROMALTSTACK = 0x6A,
         // Stack
+        DUPFROMALTSTACK = 0x6A,
         TOALTSTACK = 0x6B, // Puts the input onto the top of the alt stack. Removes it from the main stack.
         FROMALTSTACK = 0x6C, // Puts the input onto the top of the main stack. Removes it from the alt stack.
         XDROP = 0x6D,
@@ -125,12 +125,7 @@
         UNPACK = 0xC2,
         PICKITEM = 0xC3,
         SETITEM = 0xC4,
-        NEWARRAY = 0xC5,  //用作引用類型
-        //CLONE = 0xC6,
-        NEWSTRUCT = 0xC7,//用作值類型
-        //CLONESTRUCTONLY = 0xC8,//仅仅当类型是值类型时Clone， //其它do nothing。
-        //IL的值类型处理不生成特别指令，实在虚拟机内部判断何时Copy的，我们需要插入这些代码
-
-
+        NEWARRAY = 0xC5, //用作引用類型
+        NEWSTRUCT = 0xC6, //用作值類型
     }
 }
