+++ src/AntShares.VM/Types/InteropInterface.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+
 namespace AntShares.VM.Types
 {
     internal class InteropInterface : StackItem
